+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2019-02-23
+### Added
+- New command: `!rumor` - Tells a random rumor.
+- New command: `!mutation` - Draws a random mutation.
+- New command: `!arkthreat` - Draws a random threat against the Ark.
+- Bot's response to its mention, which lets him answer its prefix.
+- New JS Class: YZEmbed (extends Discord.RichEmbed with predefined properties).
+
 ## [1.1.0] - 2019-02-22
 ### Added
 - Support for negative dice rolls.